+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Descriptive changelog
 (All dates are DD.MM.YYYY)
 
-
+#### Updatable SHA hash implementation - 0.0.2 - 21.7.2019
+- Added "updatable" version for SHA
+- Moved sha and blake to hash package
+- Updated tests
 
 #### Initial release - 0.0.1 - 20.7.2019
 - Implemented Blake2b and SHA256/512